Remove the (currently) unused ICloud::Reminder
This used to be the primary interface to reminders, but I got
side-tracked with the actual implementation, which uses
ICloud::Records::Reminder (among other records). The record interface
maps closely (as far as I can tell) to iCloud's internal data
structures, but is not a very friendly or Rubyish way of creating stuff.