Add Record.has_many
This is a bit of a crappy implementation, but that can be cleaned up.
The has_many interface is fine, I think, and is necessary for other apps
(contacts in particular) to make any sense. Reminders can be refactored
to use this instead of the weird Reminder#alarms= stuff which basically
does the same thing but more ambiguously.