Mostly extract reminders app from Session
Much of the stuff which I thought was shared by all of the icloud web
apps was actually specific to reminders. Whoops. So I'm extracting it to
a separate class, and leaving Session to deal with logging in and
actually issuing the HTTP requests.

Some of this stuff might be extracted again to some sort of abstract
base class or mixin for apps, but I think it's more important that
implementing the other apps is actually possible than it is to minimize
duplication.

All the tests are failing, deal with it